Assignment 04 - WineAndFoodPair.com (initial submit)
Initial submit for “www.wineandfoodpair.com” for Assignment #4.
— Click on Any Food Icon.
— SlideToggle to show Wines that will pair w/ it.
— Click again on that Food Icon to Auto-Hide it.

//To-Do (for Final):
— Also do the bathing for Wine up to Food (not just Food to Wine)
— Check on SlideUp (Hide) whenever any buttons are clicked. Right now,
it is only slide down/up for the same button, then you have to click
the next food icon. Manual reset… We need to be smarter on how to
auto-reset on any Icon getting clicked (and checking to see if current
SECTION is shown or hidden first.